requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langchain-text-splitters
boto3
docx2txt
langchain_aws
fastapi
streamlit
python-multipart
bs4